Extract SeshatIndexManager into its own file
